# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This API documentation describes endpoints and models you will face with during interaction with KadoCore APIs",
        "title": "KadoCore API",
        "contact": {
            "name": "Back-end engineer",
            "email": "ka1tzyu@gmail.com"
        },
        "version": "1.0"
    },
    "host": "localhost:4000",
    "basePath": "/api/v1",
    "paths": {
...
        "/manufacturers": {
            "get": {
                "description": "Helps to retrieve a list of all manufacturers",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "manufacturers"
                ],
                "summary": "Get all manufacturers",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/types.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/data.GetManufacturerResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Unexpected database error",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Helps to create a new manufacturer",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "manufacturers"
                ],
                "summary": "Create a manufacturer",
                "parameters": [
                    {
                        "description": "create manufacturer body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/data.CreateManufacturerRequest"
                        }
                    }
                ],
                "responses": {
                    "401": {
                        "description": "User is not logged in",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Action is forbidden for user of this role",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Unexpected database error",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/manufacturers/{id}": {
            "get": {
                "description": "Helps to get the manufacturer with the specified id",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "manufacturers"
                ],
                "summary": "Get single manufacturer by id",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/types.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/data.GetManufacturerResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Incorrect id path",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Manufacturer not found",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Unexpected database error",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Helps to delete the existing manufacturer",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "manufacturers"
                ],
                "summary": "Delete a manufacturer",
                "responses": {
                    "400": {
                        "description": "Incorrect id path",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "User is not logged in",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Action is forbidden for user of this role",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Unexpected database error",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Helps to update the existing manufacturer",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "manufacturers"
                ],
                "summary": "Update a manufacturer",
                "parameters": [
                    {
                        "description": "update manufacturer body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/data.UpdateManufacturerRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Incorrect id path",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "User is not logged in",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Action is forbidden for user of this role",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Unexpected database error",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "data.CreateManufacturerRequest": {
            "type": "object",
            "required": [
                "original_title"
            ],
            "properties": {
                "description_eng": {
                    "type": "string",
                    "minLength": 6,
                    "example": "Chinese manufacturer that is popular for Naruto collection cards"
                },
                "description_ukr": {
                    "type": "string",
                    "minLength": 6,
                    "example": "Китайський виробник, популярний за колекційні картки по Наруто"
                },
                "image": {
                    "type": "string",
                    "example": "https://example.com/image.png"
                },
                "original_title": {
                    "type": "string",
                    "minLength": 4,
                    "example": "Kayou"
                }
            }
        },
...
        "data.GetManufacturerResponse": {
            "type": "object",
            "properties": {
                "description_eng": {
                    "type": "string",
                    "example": "Chinese manufacturer that is popular for Naruto collection cards"
                },
                "description_ukr": {
                    "type": "string",
                    "example": "Китайський виробник, популярний за колекційні картки по Наруто"
                },
                "id": {
                    "type": "string",
                    "format": "uuid",
                    "example": "3c1e3b82-3a29-4cc0-a4b2-4e7c4ac58052"
                },
                "image": {
                    "type": "string",
                    "example": "https://example.com/image.png"
                },
                "original_title": {
                    "type": "string",
                    "example": "Kayou"
                }
            }
        },
...
        "data.UpdateManufacturerRequest": {
            "type": "object",
            "properties": {
                "description_eng": {
                    "type": "string",
                    "minLength": 6,
                    "example": "Chinese manufacturer that is popular for Naruto collection cards"
                },
                "description_ukr": {
                    "type": "string",
                    "minLength": 6,
                    "example": "Китайський виробник, популярний за колекційні картки по Наруто"
                },
                "image": {
                    "type": "string",
                    "example": "https://example.com/image.png"
                },
                "original_title": {
                    "type": "string",
                    "minLength": 4,
                    "example": "Kayou"
                }
            }
        },
...
        "types.DetailedError": {
            "type": "object",
            "properties": {
                "detail": {
                    "description": "detail of field's error, if error is unknown, this field will be \"\" (empty string)",
                    "type": "string"
                },
                "field": {
                    "description": "problematic field on which error occured, if error has no specific errored field (in case of general error) this field will be \"\" (empty string)",
                    "type": "string"
                }
            }
        },
        "types.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "description": "error type",
                    "type": "string"
                },
                "errors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/types.DetailedError"
                    }
                },
                "message": {
                    "description": "messsage describing an error",
                    "type": "string"
                }
            }
        },
...
        "types.SuccessResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}